requirements.txt
asgiref==3.4.1
astroid==2.7.3
autopep8==1.6.0
bcrypt==3.2.0
black==21.8b0
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
dnspython==1.16.0
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
peewee==3.14.8
platformdirs==2.3.0
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pylint==2.10.2
pymongo==3.12.0
PyMySQL==1.0.2
python-dotenv==0.19.0
regex==2021.8.28
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing==3.7.4.3
typing-extensions==3.10.0.2
uvicorn==0.15.0
wrapt==1.12.1
